requirements.txt
Django>=5.0,<6.0
python-decouple>=3.8
requests>=2.31
django-allauth[google]>=65.0
